feat: raise compatibility to TYPO3 v12 and v13
