+++ Configuration/JavaScriptModules.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    'dependencies' => [
+        'backend'
+    ],
+    'tags' => [
+        'backend.form',
+    ],
+    'imports' => [
+        '@colorcube/dummy_content/' => 'EXT:dummy_content/Resources/Public/JavaScript/ESM/',
+    ],
+];
