+++ Configuration/TCA/Overrides/pages.php
@@ -1,8 +1,6 @@
 <?php
 
 
-defined('TYPO3_MODE') || die();
-
 (static function () {
 
     // Page titles:
