+++ Configuration/TCA/Overrides/tt_content.php
@@ -1,6 +1,4 @@
 <?php
-defined('TYPO3_MODE') || die();
-
 (static function () {
 
     // Headers in Content Elements:
