+++ Configuration/TCA/Overrides/tx_news_domain_model_news.php
@@ -1,7 +1,5 @@
 <?php
 
-defined('TYPO3_MODE') || die();
-
 (static function () {
 
     if (isset($GLOBALS['TCA']['tx_news_domain_model_news'])) {
