+++ composer.json
@@ -22,8 +22,8 @@
         }
     ],
     "require": {
-        "typo3/cms-backend": "^8.7 || ^9.5 || ^10.4 || ^11.5",
-        "typo3/cms-core": "^8.7 || ^9.5 || ^10.4 || ^11.5"
+        "typo3/cms-backend": "^12.4 || ^13.4",
+        "typo3/cms-core": "^12.4 || ^13.4"
     },
     "replace": {
         "typo3-ter/dummy-content": "self.version"
